Move Sign In navigation to navigation folder
