+++ .storybook/components/SignIn/SignIn.stories.tsx
@@ -1,11 +1,7 @@
 import { TiFAPI, createAWSTiFAPIFetch } from "@api-client"
 import { createForgotPasswordScreens } from "@auth/forgot-password"
 import { createForgotPasswordEnvironment } from "@auth/forgot-password/Environment"
-import {
-  CognitoSignInAuthenticator,
-  SignInParamsList,
-  createSignInScreens
-} from "@auth/sign-in"
+import { CognitoSignInAuthenticator } from "@auth/sign-in"
 import {
   cognitoConfirmSignUpWithAutoSignIn,
   createSignUpEnvironment
@@ -16,6 +12,10 @@ import { API_URL } from "@env"
 import { TiFQueryClientProvider } from "@lib/ReactQuery"
 import { NavigationContainer } from "@react-navigation/native"
 import { createStackNavigator } from "@react-navigation/stack"
+import {
+  SignInParamsList,
+  createSignInScreens
+} from "@root-feature/navigation/auth/SignIn"
 import { createSignUpScreens } from "@root-feature/navigation/auth/SignUp"
 import { SettingsScreen } from "@screens/SettingsScreen/SettingsScreen"
 import { ComponentMeta, ComponentStory } from "@storybook/react-native"
