+++ auth/sign-in/index.ts
@@ -1,2 +1,2 @@
 export * from "./Authenticator"
-export * from "./Navigation"
+export * from "./SignInForm"
